Unify SoundFontProvider to FluidsynthMidiService.
It would have been required (just did not want to do because now the app
project became too big) after getting existing sf2 loader stable.
So now is the time.

In the next steps... it needs to make use of this soundfont provider
feature to explicitly specify the selected sf2 for fluidsynth.
The provider feature also needs to support non-user file resources such as
assets or OBB files.